INTYGFV-14128: Fetch certificate from Intygstjanst if it doens´t exist in Webcert. Made necessary adjustmets in services dependent on the certificate existing in Webcert. In addition the patient is now always fetched from PU.
