+++ web/src/main/java/se/inera/intyg/webcert/web/service/facade/util/IntygToCertificateConverter.java
@@ -1,0 +1,9 @@
+package se.inera.intyg.webcert.web.service.facade.util;
+
+import se.inera.intyg.common.support.facade.model.Certificate;
+import se.inera.intyg.webcert.web.service.intyg.dto.IntygContentHolder;
+
+public interface IntygToCertificateConverter {
+
+    Certificate convert(IntygContentHolder intygContentHolder);
+}
