+++ web/src/main/java/se/inera/intyg/webcert/web/service/facade/util/UtkastToCertificateConverter.java
@@ -22,7 +22,7 @@ package se.inera.intyg.webcert.web.service.facade.util;
 import se.inera.intyg.common.support.facade.model.Certificate;
 import se.inera.intyg.webcert.persistence.utkast.model.Utkast;
 
-public interface CertificateConverter {
+public interface UtkastToCertificateConverter {
 
     Certificate convert(Utkast certificate);
 }
