--- web/src/test/java/se/inera/intyg/webcert/web/service/facade/util/CertificateConverterTest.java
@@ -1,777 +1,0 @@
-/*
- * Copyright (C) 2021 Inera AB (http://www.inera.se)
- *
- * This file is part of sklintyg (https://github.com/sklintyg).
- *
- * sklintyg is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * sklintyg is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package se.inera.intyg.webcert.web.service.facade.util;
-
-import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.doReturn;
-import static org.mockito.Mockito.mock;
-
-import java.time.LocalDateTime;
-import java.util.stream.Stream;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import se.inera.intyg.common.services.texts.IntygTextsService;
-import se.inera.intyg.common.support.common.enumerations.RelationKod;
-import se.inera.intyg.common.support.facade.builder.CertificateBuilder;
-import se.inera.intyg.common.support.facade.model.Certificate;
-import se.inera.intyg.common.support.facade.model.CertificateRelationType;
-import se.inera.intyg.common.support.facade.model.CertificateStatus;
-import se.inera.intyg.common.support.facade.model.Patient;
-import se.inera.intyg.common.support.facade.model.PersonId;
-import se.inera.intyg.common.support.facade.model.metadata.CertificateMetadata;
-import se.inera.intyg.common.support.facade.model.metadata.Unit;
-import se.inera.intyg.common.support.model.UtkastStatus;
-import se.inera.intyg.common.support.modules.registry.IntygModuleRegistry;
-import se.inera.intyg.common.support.modules.support.api.ModuleApi;
-import se.inera.intyg.schemas.contract.Personnummer;
-import se.inera.intyg.webcert.common.model.WebcertCertificateRelation;
-import se.inera.intyg.webcert.persistence.utkast.model.Utkast;
-import se.inera.intyg.webcert.persistence.utkast.model.VardpersonReferens;
-import se.inera.intyg.webcert.web.service.relation.CertificateRelationService;
-import se.inera.intyg.webcert.web.web.controller.api.dto.Relations;
-import se.inera.intyg.webcert.web.web.controller.api.dto.Relations.FrontendRelations;
-
-@ExtendWith(MockitoExtension.class)
-public class CertificateConverterTest {
-
-    @Mock
-    private IntygModuleRegistry moduleRegistry;
-
-    @Mock
-    private CertificateRelationService certificateRelationService;
-
-    @Mock
-    private IntygTextsService intygTextsService;
-
-    @Mock
-    private PatientConverter patientConverter;
-
-    @InjectMocks
-    private CertificateConverterImpl certificateConverter;
-
-    private final Utkast draft = createDraft();
-    private final Relations relations = new Relations();
-    private final Patient patient = getPatient();
-
-    @BeforeEach
-    void setupMocks() throws Exception {
-        final var moduleApi = mock(ModuleApi.class);
-        doReturn(moduleApi)
-            .when(moduleRegistry).getModuleApi(draft.getIntygsTyp(), draft.getIntygTypeVersion());
-
-        doReturn(createCertificate())
-            .when(moduleApi).getCertificateFromJson(draft.getModel());
-
-        doReturn(relations)
-            .when(certificateRelationService).getRelations(draft.getIntygsId());
-
-        doReturn(patient)
-            .when(patientConverter).convert(draft);
-    }
-
-    @Nested
-    class ValidateCommonMetadata {
-
-        @Test
-        void shallIncludeCreatedDateTime() {
-            final var expectedCreated = draft.getSkapad();
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedCreated, actualCertificate.getMetadata().getCreated());
-        }
-
-        @ParameterizedTest
-        @ValueSource(ints = {0, 1, 100})
-        void shallIncludeVersion(int expectedVersion) {
-            draft.setVersion(expectedVersion);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedVersion, actualCertificate.getMetadata().getVersion());
-        }
-
-        @ParameterizedTest
-        @ValueSource(booleans = {true, false})
-        void shallIncludeForwarded(boolean expectedForwarded) {
-            draft.setVidarebefordrad(expectedForwarded);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedForwarded, actualCertificate.getMetadata().isForwarded());
-        }
-
-        @ParameterizedTest
-        @ValueSource(booleans = {true, false})
-        void shallIncludeTestCertificate(boolean expectedTestCertificate) {
-            draft.setTestIntyg(expectedTestCertificate);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedTestCertificate, actualCertificate.getMetadata().isTestCertificate());
-        }
-
-        @ParameterizedTest
-        @ValueSource(booleans = {true, false})
-        void shallIncludeLatestMajorVersion(boolean expectedLatestMajorVersion) {
-            doReturn(expectedLatestMajorVersion).when(intygTextsService).isLatestMajorVersion(any(), any());
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedLatestMajorVersion, actualCertificate.getMetadata().isLatestMajorVersion());
-        }
-    }
-
-    @Nested
-    class ValidateUnit {
-
-        @Test
-        void shallContainCompleteUnitData() {
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertAll(
-                () -> assertNotNull(actualCertificate.getMetadata().getUnit().getUnitId(), "UnitId should not be null"),
-                () -> assertNotNull(actualCertificate.getMetadata().getUnit().getUnitName(), "UnitName should not be null"),
-                () -> assertNotNull(actualCertificate.getMetadata().getUnit().getAddress(), "Address should not be null"),
-                () -> assertNotNull(actualCertificate.getMetadata().getUnit().getZipCode(), "ZipCode should not be null"),
-                () -> assertNotNull(actualCertificate.getMetadata().getUnit().getCity(), "City should not be null"),
-                () -> assertNotNull(actualCertificate.getMetadata().getUnit().getEmail(), "Email should not be null"),
-                () -> assertNotNull(actualCertificate.getMetadata().getUnit().getPhoneNumber(), "Phonenumber should not be null")
-            );
-        }
-    }
-
-    @Nested
-    class ValidateCareProvider {
-
-        @Test
-        void shallIncludeCareProviderId() {
-            final var expectedCareProviderId = "CareProviderId";
-            draft.setVardgivarId(expectedCareProviderId);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedCareProviderId, actualCertificate.getMetadata().getCareProvider().getUnitId());
-        }
-
-        @Test
-        void shallIncludeCareProviderName() {
-            final var expectedCareProviderName = "CareProviderName";
-            draft.setVardgivarNamn(expectedCareProviderName);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedCareProviderName, actualCertificate.getMetadata().getCareProvider().getUnitName());
-        }
-    }
-
-    @Nested
-    class ValidateIssuedBy {
-
-        @Test
-        void shallIncludePersonId() {
-            final var expectedPersonId = "PersonId";
-            draft.getSkapadAv().setHsaId(expectedPersonId);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedPersonId, actualCertificate.getMetadata().getIssuedBy().getPersonId());
-        }
-
-        @Test
-        void shallIncludeName() {
-            final var expectedFullName = "Doctor Alpha";
-            draft.getSkapadAv().setNamn(expectedFullName);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedFullName, actualCertificate.getMetadata().getIssuedBy().getFullName());
-        }
-    }
-
-
-    @Test
-    void shallIncludePatient() {
-        final var expectedPatient = getPatient();
-        final var actualCertificate = certificateConverter.convert(draft);
-        assertEquals(expectedPatient, actualCertificate.getMetadata().getPatient());
-    }
-
-    @Nested
-    class ValidateStatus {
-
-        @Test
-        void shallIncludeStatusUnsigned() {
-            final var expectedStatus = CertificateStatus.UNSIGNED;
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getStatus());
-        }
-
-        @Test
-        void shallIncludeStatusSigned() {
-            final var expectedStatus = CertificateStatus.SIGNED;
-            draft.setStatus(UtkastStatus.SIGNED);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getStatus());
-        }
-
-        @Test
-        void shallIncludeStatusRevoked() {
-            final var expectedStatus = CertificateStatus.REVOKED;
-            draft.setStatus(UtkastStatus.SIGNED);
-            draft.setAterkalladDatum(LocalDateTime.now());
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getStatus());
-        }
-
-        @Test
-        void shallIncludeStatusLocked() {
-            final var expectedStatus = CertificateStatus.LOCKED;
-            draft.setStatus(UtkastStatus.DRAFT_LOCKED);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getStatus());
-        }
-
-        @Test
-        void shallIncludeStatusLockedRevoked() {
-            final var expectedStatus = CertificateStatus.LOCKED_REVOKED;
-            draft.setStatus(UtkastStatus.DRAFT_LOCKED);
-            draft.setAterkalladDatum(LocalDateTime.now());
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getStatus());
-        }
-    }
-
-    @Nested
-    class ValidateParentRelation {
-
-        private WebcertCertificateRelation parentRelation;
-
-        @BeforeEach
-        void setup() {
-            parentRelation = new WebcertCertificateRelation(
-                "ParentId",
-                RelationKod.ERSATT,
-                LocalDateTime.now(),
-                UtkastStatus.SIGNED,
-                false
-            );
-
-            relations.setParent(parentRelation);
-        }
-
-        @Test
-        void shallNotIncludeParent() {
-            relations.setParent(null);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertNull(actualCertificate.getMetadata().getRelations().getParent());
-        }
-
-        @Test
-        void shallIncludeParentRelationCertificateId() {
-            final var expectedParentId = "ParentId";
-            parentRelation.setIntygsId(expectedParentId);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedParentId, actualCertificate.getMetadata().getRelations().getParent().getCertificateId());
-        }
-
-        @Test
-        void shallIncludeParentRelationCreateDateTime() {
-            final var expectedCreatedDateTime = LocalDateTime.now();
-            parentRelation.setSkapad(expectedCreatedDateTime);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedCreatedDateTime, actualCertificate.getMetadata().getRelations().getParent().getCreated());
-        }
-
-        @Test
-        void shallIncludeParentRelationStatusUnsignedWhenComplete() {
-            final var expectedStatus = CertificateStatus.UNSIGNED;
-            parentRelation.setStatus(UtkastStatus.DRAFT_COMPLETE);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getRelations().getParent().getStatus());
-        }
-
-        @Test
-        void shallIncludeParentRelationStatusUnsignedWhenIncomplete() {
-            final var expectedStatus = CertificateStatus.UNSIGNED;
-            parentRelation.setStatus(UtkastStatus.DRAFT_INCOMPLETE);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getRelations().getParent().getStatus());
-        }
-
-        @Test
-        void shallIncludeParentRelationStatusSigned() {
-            final var expectedStatus = CertificateStatus.SIGNED;
-            parentRelation.setStatus(UtkastStatus.SIGNED);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getRelations().getParent().getStatus());
-        }
-
-        @Test
-        void shallIncludeParentRelationStatusRevoked() {
-            final var expectedStatus = CertificateStatus.REVOKED;
-            parentRelation.setStatus(UtkastStatus.SIGNED);
-            parentRelation.setMakulerat(true);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getRelations().getParent().getStatus());
-        }
-
-        @Test
-        void shallIncludeParentRelationStatusLocked() {
-            final var expectedStatus = CertificateStatus.LOCKED;
-            parentRelation.setStatus(UtkastStatus.DRAFT_LOCKED);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getRelations().getParent().getStatus());
-        }
-
-        @Test
-        void shallIncludeParentRelationStatusLockedRevoked() {
-            final var expectedStatus = CertificateStatus.LOCKED_REVOKED;
-            parentRelation.setStatus(UtkastStatus.DRAFT_LOCKED);
-            parentRelation.setMakulerat(true);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getRelations().getParent().getStatus());
-        }
-
-        @Test
-        void shallIncludeParentRelationReplaced() {
-            final var expectedRelationsType = CertificateRelationType.REPLACED;
-            parentRelation.setRelationKod(RelationKod.ERSATT);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedRelationsType, actualCertificate.getMetadata().getRelations().getParent().getType());
-        }
-
-        @Test
-        void shallIncludeParentRelationCopied() {
-            final var expectedRelationsType = CertificateRelationType.COPIED;
-            parentRelation.setRelationKod(RelationKod.KOPIA);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedRelationsType, actualCertificate.getMetadata().getRelations().getParent().getType());
-        }
-
-        @Test
-        void shallIncludeParentRelationExtended() {
-            final var expectedRelationsType = CertificateRelationType.EXTENDED;
-            parentRelation.setRelationKod(RelationKod.FRLANG);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedRelationsType, actualCertificate.getMetadata().getRelations().getParent().getType());
-        }
-
-        @Test
-        void shallIncludeParentRelationComplemented() {
-            final var expectedRelationsType = CertificateRelationType.COMPLEMENTED;
-            parentRelation.setRelationKod(RelationKod.KOMPLT);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedRelationsType, actualCertificate.getMetadata().getRelations().getParent().getType());
-        }
-    }
-
-    @Nested
-    @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-    class ValidateChildRelation {
-
-        @BeforeEach
-        void setup() {
-            final var parentRelation = new WebcertCertificateRelation(
-                "ParentId",
-                RelationKod.ERSATT,
-                LocalDateTime.now(),
-                UtkastStatus.SIGNED,
-                false
-            );
-
-            relations.setParent(parentRelation);
-        }
-
-        Stream<FrontendRelations> allChildRelations() {
-            return Stream.of(
-                getReplacedByDraft().getLatestChildRelations(),
-                getReplacedByCertificate().getLatestChildRelations(),
-                getComplementedByDraft().getLatestChildRelations(),
-                getComplementedByCertificate().getLatestChildRelations(),
-                getDraftCopy().getLatestChildRelations()
-            );
-        }
-
-        Stream<FrontendRelations> complementedChildRelations() {
-            return Stream.of(
-                getComplementedByDraft().getLatestChildRelations(),
-                getComplementedByCertificate().getLatestChildRelations()
-            );
-        }
-
-        Stream<FrontendRelations> replacedChildRelations() {
-            return Stream.of(
-                getReplacedByDraft().getLatestChildRelations(),
-                getReplacedByCertificate().getLatestChildRelations()
-            );
-        }
-
-        Stream<FrontendRelations> copiedChildRelations() {
-            return Stream.of(
-                getDraftCopy().getLatestChildRelations()
-            );
-        }
-
-        @Test
-        void shallNotIncludeParent() {
-            relations.setParent(null);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertNull(actualCertificate.getMetadata().getRelations().getParent());
-        }
-
-        @ParameterizedTest
-        @MethodSource("allChildRelations")
-        void shallIncludeChildRelationCertificateId(FrontendRelations childRelations) {
-            final var childRelation = getChildRelationToTest(childRelations);
-            final var expectedCertificateId = childRelation.getIntygsId();
-            childRelation.setIntygsId(expectedCertificateId);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedCertificateId, actualCertificate.getMetadata().getRelations().getChildren()[0].getCertificateId());
-        }
-
-        @ParameterizedTest
-        @MethodSource("allChildRelations")
-        void shallIncludeChildRelationCreateDateTime(FrontendRelations childRelations) {
-            final var expectedCreatedDateTime = LocalDateTime.now();
-
-            final var childRelation = getChildRelationToTest(childRelations);
-            childRelation.setSkapad(expectedCreatedDateTime);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedCreatedDateTime, actualCertificate.getMetadata().getRelations().getChildren()[0].getCreated());
-        }
-
-        @ParameterizedTest
-        @MethodSource("allChildRelations")
-        void shallIncludeChildRelationStatusUnsignedWhenComplete(FrontendRelations childRelations) {
-            final var expectedStatus = CertificateStatus.UNSIGNED;
-
-            final var childRelation = getChildRelationToTest(childRelations);
-            childRelation.setStatus(UtkastStatus.DRAFT_COMPLETE);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getRelations().getChildren()[0].getStatus());
-        }
-
-        @ParameterizedTest
-        @MethodSource("allChildRelations")
-        void shallIncludeChildRelationStatusUnsignedWhenIncomplete(FrontendRelations childRelations) {
-            final var expectedStatus = CertificateStatus.UNSIGNED;
-
-            final var childRelation = getChildRelationToTest(childRelations);
-            childRelation.setStatus(UtkastStatus.DRAFT_INCOMPLETE);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getRelations().getChildren()[0].getStatus());
-        }
-
-        @ParameterizedTest
-        @MethodSource("allChildRelations")
-        void shallIncludeChildRelationStatusSigned(FrontendRelations childRelations) {
-            final var expectedStatus = CertificateStatus.SIGNED;
-
-            final var childRelation = getChildRelationToTest(childRelations);
-            childRelation.setStatus(UtkastStatus.SIGNED);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getRelations().getChildren()[0].getStatus());
-        }
-
-        @ParameterizedTest
-        @MethodSource("allChildRelations")
-        void shallIncludeChildRelationStatusRevoked(FrontendRelations childRelations) {
-            final var expectedStatus = CertificateStatus.REVOKED;
-
-            final var childRelation = getChildRelationToTest(childRelations);
-            childRelation.setStatus(UtkastStatus.SIGNED);
-            childRelation.setMakulerat(true);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getRelations().getChildren()[0].getStatus());
-        }
-
-        @ParameterizedTest
-        @MethodSource("allChildRelations")
-        void shallIncludeChildRelationStatusLocked(FrontendRelations childRelations) {
-            final var expectedStatus = CertificateStatus.LOCKED;
-
-            final var childRelation = getChildRelationToTest(childRelations);
-            childRelation.setStatus(UtkastStatus.DRAFT_LOCKED);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getRelations().getChildren()[0].getStatus());
-        }
-
-        @ParameterizedTest
-        @MethodSource("allChildRelations")
-        void shallIncludeChildRelationStatusLockedRevoked(FrontendRelations childRelations) {
-            final var expectedStatus = CertificateStatus.LOCKED_REVOKED;
-
-            final var childRelation = getChildRelationToTest(childRelations);
-            childRelation.setStatus(UtkastStatus.DRAFT_LOCKED);
-            childRelation.setMakulerat(true);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedStatus, actualCertificate.getMetadata().getRelations().getChildren()[0].getStatus());
-        }
-
-        @ParameterizedTest
-        @MethodSource("replacedChildRelations")
-        void shallIncludeChildRelationReplaced(FrontendRelations childRelations) {
-            final var expectedRelationsType = CertificateRelationType.REPLACED;
-
-            final var childRelation = getChildRelationToTest(childRelations);
-            childRelation.setRelationKod(RelationKod.ERSATT);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedRelationsType, actualCertificate.getMetadata().getRelations().getChildren()[0].getType());
-        }
-
-        @ParameterizedTest
-        @MethodSource("copiedChildRelations")
-        void shallIncludeChildRelationCopied(FrontendRelations childRelations) {
-            final var expectedRelationsType = CertificateRelationType.COPIED;
-
-            final var childRelation = getChildRelationToTest(childRelations);
-            childRelation.setRelationKod(RelationKod.KOPIA);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedRelationsType, actualCertificate.getMetadata().getRelations().getChildren()[0].getType());
-        }
-
-        @ParameterizedTest
-        @MethodSource("complementedChildRelations")
-        void shallIncludeChildRelationComplemented(FrontendRelations childRelations) {
-            final var expectedRelationsType = CertificateRelationType.COMPLEMENTED;
-
-            final var childRelation = getChildRelationToTest(childRelations);
-            childRelation.setRelationKod(RelationKod.KOMPLT);
-
-            final var actualCertificate = certificateConverter.convert(draft);
-
-            assertEquals(expectedRelationsType, actualCertificate.getMetadata().getRelations().getChildren()[0].getType());
-        }
-
-        private WebcertCertificateRelation getChildRelationToTest(FrontendRelations childRelations) {
-            relations.setLatestChildRelations(childRelations);
-            if (childRelations.getReplacedByUtkast() != null) {
-                return childRelations.getReplacedByUtkast();
-            }
-            if (childRelations.getReplacedByIntyg() != null) {
-                return childRelations.getReplacedByIntyg();
-            }
-            if (childRelations.getComplementedByUtkast() != null) {
-                return childRelations.getComplementedByUtkast();
-            }
-            if (childRelations.getComplementedByIntyg() != null) {
-                return childRelations.getComplementedByIntyg();
-            }
-            if (childRelations.getUtkastCopy() != null) {
-                return childRelations.getUtkastCopy();
-            }
-            throw new IllegalArgumentException("Incorrect test data! Please verify!");
-        }
-
-        private Relations getReplacedByDraft() {
-            final var replacedByDraft = new Relations();
-            replacedByDraft.getLatestChildRelations().setReplacedByUtkast(
-                new WebcertCertificateRelation(
-                    null,
-                    RelationKod.ERSATT,
-                    LocalDateTime.now(),
-                    UtkastStatus.SIGNED,
-                    false
-                )
-            );
-            return replacedByDraft;
-        }
-
-        private Relations getReplacedByCertificate() {
-            final var replacedByIntyg = new Relations();
-            replacedByIntyg.getLatestChildRelations().setReplacedByIntyg(
-                new WebcertCertificateRelation(
-                    null,
-                    RelationKod.ERSATT,
-                    LocalDateTime.now(),
-                    UtkastStatus.SIGNED,
-                    false
-                )
-            );
-            return replacedByIntyg;
-        }
-
-        private Relations getComplementedByDraft() {
-            final var complementedByDraft = new Relations();
-            complementedByDraft.getLatestChildRelations().setComplementedByUtkast(
-                new WebcertCertificateRelation(
-                    null,
-                    RelationKod.KOMPLT,
-                    LocalDateTime.now(),
-                    UtkastStatus.SIGNED,
-                    false
-                )
-            );
-            return complementedByDraft;
-        }
-
-        private Relations getComplementedByCertificate() {
-            final var complementedByIntyg = new Relations();
-            complementedByIntyg.getLatestChildRelations().setComplementedByIntyg(
-                new WebcertCertificateRelation(
-                    null,
-                    RelationKod.KOMPLT,
-                    LocalDateTime.now(),
-                    UtkastStatus.SIGNED,
-                    false
-                )
-            );
-            return complementedByIntyg;
-        }
-
-        private Relations getDraftCopy() {
-            final var draftCopy = new Relations();
-            draftCopy.getLatestChildRelations().setUtkastCopy(
-                new WebcertCertificateRelation(
-                    null,
-                    RelationKod.KOPIA,
-                    LocalDateTime.now(),
-                    UtkastStatus.SIGNED,
-                    false
-                )
-            );
-            return draftCopy;
-        }
-    }
-
-    private Utkast createDraft() {
-        final var draft = new Utkast();
-        draft.setIntygsId("certificateId");
-        draft.setIntygsTyp("certificateType");
-        draft.setIntygTypeVersion("certificateTypeVersion");
-        draft.setModel("draftJson");
-        draft.setStatus(UtkastStatus.DRAFT_INCOMPLETE);
-        draft.setSkapad(LocalDateTime.now());
-        draft.setPatientPersonnummer(Personnummer.createPersonnummer("191212121212").orElseThrow());
-        draft.setSkapadAv(new VardpersonReferens("personId", "personName"));
-        return draft;
-    }
-
-    private Certificate createCertificate() {
-        return CertificateBuilder.create()
-            .metadata(
-                CertificateMetadata.builder()
-                    .id("certificateId")
-                    .type("certificateType")
-                    .typeVersion("certificateTypeVersion")
-                    .unit(
-                        Unit.builder()
-                            .unitId("unitId")
-                            .unitName("unitName")
-                            .address("address")
-                            .zipCode("zipCode")
-                            .city("city")
-                            .email("email")
-                            .phoneNumber("phoneNumber")
-                            .build()
-                    )
-                    .build()
-            )
-            .build();
-    }
-
-    private Patient getPatient() {
-        return Patient.builder()
-            .personId(getPersonId())
-            .firstName("Fornamnet")
-            .lastName("Efternamnet")
-            .middleName("Mellannamnet")
-            .build();
-    }
-
-    private PersonId getPersonId() {
-        final var expectedPersonId = PersonId.builder()
-            .id(draft.getPatientPersonnummer().getPersonnummer())
-            .type("PERSON_NUMMER")
-            .build();
-        return expectedPersonId;
-    }
-}
